move hooks to pyqtgraph
