+++ lavuelib/external/pyqtgraph_0_10.py
